+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Mapbox welcomes participation and contributions from everyone.
 
 # main
 
+# 11.12.3
+## Features ✨ and improvements 🏁
+* Improve the performance of `MapboxMap.cameraForCoordinates(...)` for large amounts of points.
+
 # 11.12.2 May 23, 2025
 ## Bug fixes 🐞
 * Limit view annotations collision grid size.